+++ src/index_versions/v2/mod.rs
@@ -1,6 +1,5 @@
 pub mod scores;
 pub mod search;
-pub mod stopwords;
 pub mod structs;
 
 pub const VERSION_STRING: &str = "stork-2";
