+++ src/index_versions/v3/mod.rs
@@ -1,7 +1,6 @@
 pub mod builder;
 pub mod scores;
 pub mod search;
-pub mod stopwords;
 pub mod structs;
 pub mod word_list_generators;
 pub mod writer;
