--- src/searcher/test.rs
@@ -1,1 +1,0 @@
-
